requirements.txt
# TWP Mobile - Appium Robot Framework Tests
# Requirements file for Python dependencies
# Install with: pip install -r requirements.txt

# Core Robot Framework
robotframework==7.3.2

# Appium integration for Robot Framework
robotframework-appiumlibrary==3.1

# Appium Python client
appium-python-client==5.2.4

# Selenium WebDriver (required by Appium)
selenium==4.36.0

# HTTP requests for environment checks
requests==2.32.5

# Additional utilities (installed automatically)
urllib3>=2.0.0
certifi>=2024.0.0